# Update lobsters benchmark to Rails 8.0