Fix: https://github.com/Shopify/yjit-bench/pull/358

Before:
```
inline_code_size:    6,752,176
outlined_code_size:  2,067,600
code_region_size:    17,268,736
yjit_alloc_size:     25,934,245
yjit_compile_time:    989.69ms
RSS: 393.7MiB
MAXRSS: 393.9MiB
Average of last 15, non-warmup iters: 299ms
Total time spent benchmarking: 27s

interp: ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin23]
yjit: ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]

--------  -----------  ----------  ---------  ----------  ------------  -----------
bench     interp (ms)  stddev (%)  yjit (ms)  stddev (%)  yjit 1st itr  interp/yjit
lobsters  463.9        1.3         299.8      2.6         0.563         1.547
--------  -----------  ----------  ---------  ----------  ------------  -----------
```

After:
```
inline_code_size:    6,808,680
outlined_code_size:  2,178,968
code_region_size:    17,448,960
yjit_alloc_size:     26,006,453
yjit_compile_time:    978.48ms
RSS: 380.1MiB
MAXRSS: 380.3MiB
Average of last 16, non-warmup iters: 296ms
Total time spent benchmarking: 27s

interp: ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin23]
yjit: ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]

--------  -----------  ----------  ---------  ----------  ------------  -----------
bench     interp (ms)  stddev (%)  yjit (ms)  stddev (%)  yjit 1st itr  interp/yjit
lobsters  463.1        1.8         296.3      2.2         0.596         1.563
--------  -----------  ----------  ---------  ----------  ------------  -----------
```

cc @maximecb @XrXr